+++ src/main/kotlin/org/harelang/Bundle.kt
@@ -1,13 +1,13 @@
-package org.jetbrains.plugins.template
+package org.harelang
 
 import com.intellij.DynamicBundle
 import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.Bundle"
 
-object MyBundle : DynamicBundle(BUNDLE) {
+object Bundle : DynamicBundle(BUNDLE) {
 
     @Suppress("SpreadOperator")
     @JvmStatic